+++ src/controllers/mission.controller.js
@@ -1,6 +1,7 @@
 import { StatusCodes } from "http-status-codes";
 import { addMission } from "../services/mission.service.js";
 import { bodyToMission } from "../dtos/mission.dto.js";
+import { listStoreMissions } from "../services/mission.service.js";
 
 export const handleAddMission = async (req, res, next) => {
   /*
@@ -95,3 +96,46 @@ export const handleAddMission = async (req, res, next) => {
     res.status(StatusCodes.OK).success(mission)
   };
 
+  export const handleListStoreMissions = async (req, res, next) => {
+    /*
+      #swagger.summary = '식당별 미션 목록 조회 API';
+      #swagger.responses[200] = {
+        description: "식당별 미션 목록 조회 성공 응답",
+        content: {
+          "application/json": {
+            schema: {
+              type: "object",
+              properties: {
+                resultType: { type: "string", example: "SUCCESS" },
+                error: { type: "object", nullable: true, example: null },
+                success: {
+                  type: "object",
+                  properties: {
+                    data: {
+                      type: "array",
+                      items: {
+                        type: "object",
+                        properties: {
+                          id: { type: "number" },
+                          description: { type: "string" },
+                          restaurantId: { type: "number" },
+                          score: { type: "number"
+                        }
+                      }
+                    },
+                    pagination: { type: "object", properties: { cursor: { type: "number", nullable: true } }}
+                  }
+                }
+              }
+            }
+          }
+        }
+      };
+    */
+  
+    const missions = await listStoreMissions(
+      parseInt(req.params.storeId),
+      typeof req.query.cursor === "string" ? parseInt(req.query.cursor) : 0
+    );
+    res.status(StatusCodes.OK).success(missions);
+  };
