+++ src/controllers/review.controller.js
@@ -1,6 +1,7 @@
 import { StatusCodes } from "http-status-codes";
 import { addReview } from "../services/review.service.js";
 import { bodyToReview } from "../dtos/review.dto.js";
+import { listStoreReviews } from "../services/review.service.js";
 
 export const handleAddReview = async (req, res, next) => {
   /*
@@ -74,3 +75,46 @@ export const handleAddReview = async (req, res, next) => {
   res.status(StatusCodes.OK).success(review);
 };
 
+export const handleListStoreReviews = async (req, res, next) => {
+  /*
+    #swagger.summary = '식당별 리뷰 목록 조회 API';
+    #swagger.responses[200] = {
+      description: "식당별 리뷰 목록 조회 성공 응답",
+      content: {
+        "application/json": {
+          schema: {
+            type: "object",
+            properties: {
+              resultType: { type: "string", example: "SUCCESS" },
+              error: { type: "object", nullable: true, example: null },
+              success: {
+                type: "object",
+                properties: {
+                  data: {
+                    type: "array",
+                    items: {
+                      type: "object",
+                      properties: {
+                        id: { type: "number" },
+                        store: { type: "object", properties: { id: { type: "number" }, name: { type: "string" } } },
+                        user: { type: "object", properties: { id: { type: "number" }, email: { type: "string" }, name: { type: "string" } } },
+                        content: { type: "string" }
+                      }
+                    }
+                  },
+                  pagination: { type: "object", properties: { cursor: { type: "number", nullable: true } }}
+                }
+              }
+            }
+          }
+        }
+      }
+    };
+  */
+  const reviews = await listStoreReviews(
+    parseInt(req.params.storeId),
+    typeof req.query.cursor === "string" ? parseInt(req.query.cursor) : 0
+  );
+
+  res.status(StatusCodes.OK).success(reviews);
+};
