+++ src/repositories/mission.repository.js
@@ -28,3 +28,14 @@ export const checkMissionExists = async (data) => {
   });
   return count > 0;
 }
+
+export const getAllStoreMissions = async (restaurantId, cursor) => {
+  const missions = await prisma.mission.findMany({
+    select: { id: true, description: true, restaurantId: true, score: true },
+    where: { restaurantId: restaurantId, id: { gt: cursor } },
+    orderBy: { id: "asc" },
+    take: 3,
+  });
+
+  return missions;
+};
