+++ src/repositories/user.repository.js
@@ -44,16 +44,6 @@ export const getUserPreferencesByUserId = async (userId) => {
   return preferences;
 };
 
-export const getAllStoreReviews = async (restaurantId, cursor) => {
-  const reviews = await prisma.review.findMany({
-    select: { id: true, content: true, restaurantId: true, userId: true },
-    where: { restaurantId: restaurantId, id: { gt: cursor } },
-    orderBy: { id: "asc" },
-    take: 3,
-  });
-
-  return reviews;
-}; 
 
 export const getAllUserReviews = async (userId, cursor) => {
   const reviews = await prisma.review.findMany({
@@ -65,13 +55,3 @@ export const getAllUserReviews = async (userId, cursor) => {
   return reviews;
 };
 
-export const getAllStoreMissions = async (restaurantId, cursor) => {
-  const missions = await prisma.mission.findMany({
-    select: { id: true, description: true, restaurantId: true, score: true },
-    where: { restaurantId: restaurantId, id: { gt: cursor } },
-    orderBy: { id: "asc" },
-    take: 3,
-  });
-
-  return missions;
-};
