+++ src/services/restaurant.service.js
@@ -1,10 +1,7 @@
 import { createRestaurant } from '../repositories/restaurant.repository.js';
-import { getAllStoreReviews, getAllStoreMissions } from '../repositories/user.repository.js';
 import { responseFromRestaurant } from '../dtos/restaurant.dto.js';
-import { responseFromReviews } from '../dtos/review.dto.js';
-import { responseFromMissions } from '../dtos/mission.dto.js';
 import { getRestaurant, checkRestaurantExists } from '../repositories/restaurant.repository.js';
-import { ResataurantAlreadyExistError, RestaurantNotExistError } from '../errors.js';
+import { ResataurantAlreadyExistError } from '../errors.js';
 
 export const addRestaurant = async (data) => {
     const restaurantId = await createRestaurant({ // restaurantId가 아니라 restaurantData로 수정해야 할 듯
@@ -21,20 +18,5 @@ export const addRestaurant = async (data) => {
     return responseFromRestaurant(restaurant)
   }
 
-export const listStoreReviews = async (restaurantId, cursor = 0) => {
-  const exists = await checkRestaurantExists(restaurantId);
-  if (!exists) {
-    throw new RestaurantNotExistError("리뷰를 조회하려는 가게가 존재하지 않습니다.", restaurantId);
-  }
-  const reviews = await getAllStoreReviews(restaurantId,cursor);
-  return responseFromReviews(reviews);
-};
 
-export const listStoreMissions = async (restaurantId, cursor = 0) => {
-  const exists = await checkRestaurantExists(restaurantId);
-  if (!exists) {
-    throw new RestaurantNotExistError("미션을 조회하려는 가게가 존재하지 않습니다.", restaurantId);
-  }
-  const missions = await getAllStoreMissions(restaurantId,cursor);
-  return responseFromMissions(missions);
-};
+
