+++ src/services/review.service.js
@@ -2,6 +2,7 @@ import { createReview, getReview } from '../repositories/review.repository.js';
 import { checkRestaurantExists } from '../repositories/restaurant.repository.js';
 import { responseFromReviews } from '../dtos/review.dto.js';
 import { RestaurantNotExistError } from '../errors.js';
+import { getAllStoreReviews } from '../repositories/review.repository.js';
 
 export const addReview = async (data) => {
   const exists = await checkRestaurantExists(data.restaurantId);
@@ -19,3 +20,11 @@ export const addReview = async (data) => {
   return responseFromReviews(review);
 };
 
+export const listStoreReviews = async (restaurantId, cursor = 0) => {
+  const exists = await checkRestaurantExists(restaurantId);
+  if (!exists) {
+    throw new RestaurantNotExistError("리뷰를 조회하려는 가게가 존재하지 않습니다.", restaurantId);
+  }
+  const reviews = await getAllStoreReviews(restaurantId,cursor);
+  return responseFromReviews(reviews);
+};
